finalização update de registros

+++ app/Services/googleDrive/googledrive.ts
@@ -233,11 +233,18 @@ async function searchFile(authClient, fileName, parentId = undefined) {
 }
 
 async function deleteFile(authClient, fileId) {
-  const drive = google.drive({ version: 'v3', auth: authClient });
-  const request = drive.files.delete({
-    'fileId': fileId
-  })
-  return request
+  try {
+    const drive = google.drive({ version: 'v3', auth: authClient });
+    if (fileId) {
+      const request = drive.files.delete({
+        'fileId': fileId
+      })
+      return request
+    }
+  } catch (error) {
+    throw error
+  }
+
 }
 
 //RENOMERAR ARQUIVOS**************************************************************** */
